Move converters and enums
Move it in different folders in GameEndPoint
